ref: move remaining tests in message.rs
