+++ src/tests/conformance/mod.rs
@@ -1,6 +1,8 @@
 mod basic_query;
+mod get_blocks;
+mod get_data;
+mod get_header;
 mod handshake;
 mod invalid_message;
-mod messages;
 mod peering;
 mod unsolicited_response;
